requirements.txt
Django==1.11
Jinja2==2.8
Markdown==2.6.2
MarkupSafe==0.23
Pillow>2.8.2
PyYAML==3.11
Pygments==2.0.2
Unidecode==0.04.18
Willow>0.2.2
ansible==2.0.1.0
beautifulsoup4==4.4.0
django-appconf==1.0.1
django-compressor==1.5
django-medusa==0.3.0
django-modelcluster==1.1
django-sendfile==0.3.10
django-taggit==0.18.0
django-treebeard==3.0
djangorestframework==3.7
ecdsa==0.13
elasticsearch==1.7.0
google-api-python-client==1.5.0
html5lib==0.999
httplib2==0.9.2
oauth2client==2
paramiko==1.16.0
postgres==2.1.2
psycopg2==2.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
python-memcached==1.57
pytz==2015.4
requests==2.7.0
rsa==3.3
simplejson==3.8.2
six==1.9.0
uritemplate==0.6
urllib3==1.12
wagalytics==0.7
wagtail==2.0
wagtailfontawesome==1.1.3
wheel==0.24.0
